move tests with users to with-users directory - replace ignoreTestFiles with spec path
