+++ integration/cypress/integration/with-users/base/footer.spec.ts
