+++ integration/cypress/integration/with-users/devices/list.spec.ts
