+++ integration/cypress/integration/with-users/domains/list.spec.ts
