+++ integration/cypress/integration/with-users/machines/list.spec.ts
