+++ integration/cypress/integration/with-users/preferences/base.spec.ts
