+++ integration/cypress/integration/with-users/zones/list.spec.ts
